app/requirements.txt
requests==2.24.0
pandas==1.1.3
numpy
attrs==19.3.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
google-api-core==1.22.2
google-auth==1.21.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.1
google-cloud-pubsub==2.1.0
google-cloud-storage==1.31.2
google-cloud-trace==0.23.0
google-cloud==0.34.0
google-cloud-secret-manager==2.0.0
grpcio==1.30.0
idna==2.9
Jinja2==2.11.2
MarkupSafe==1.1.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
urllib3==1.25.9
tensorflow==1.15.0
scipy
flask
uwsgi
opencv-python==3.4.5.20
mock
pytest